Ajout de la deuxième page : Affichage du prix estimé

+++ Application/app/src/main/java/com/example/miniprojet/backend/ModeleInterface.kt
@@ -2,5 +2,5 @@ package com.example.miniprojet.backend
 
 interface ModeleInterface {
 
-    fun priceCalculator(surfaceTerrain: Int, surfaceBatiment: Int, nbPiece: Int, longitude: Int, latitude: Int): Int
+    fun priceCalculator(surfaceTerrain: Int, surfaceBatiment: Int, nbPiece: Int, longitude: Int, latitude: Int): Double
 }
